:sparkles: feat : [inception] Finish FTP server with wordpress

+++ inception/srcs/requirements/bonus/ftp/tools/entrypoint.sh
@@ -1,10 +1,16 @@
 #!/usr/bin/env sh
 
-set -x
+set -e
 
-cp /tmp/vsftpd.conf /etc/vsftpd/vsftpd.conf
-chown -R vsftp:ftp /var/www
-mkdir /home/vsftp && chown -R vsftp:ftp /home/vsftp
+if [ ! -f .setup ]; then
+	echo "Setting up vsftpd"
+	cp /tmp/vsftpd.conf /etc/vsftpd/vsftpd.conf
+	chown -R www-data:www-data /var/www/wordpress
+	echo -e "$FTP_PASSWORD\n$FTP_PASSWORD" | passwd www-data
+	touch .setup
+else
+	echo "vsftpd is already setup"
+fi
 
 echo "Execute $@"
 exec "$@"
